config: split the 3,120-line config.go per domain
config.go held the entire product's config schema, its loading/path
resolution, AND a hand-rolled validation framework (sectionValidators +
~25 validateX helpers) in one file. Separate concerns into focused
same-package files:

  - config_validate.go — the sectionValidator framework and every
    validateX helper (the validation logic).
  - config_peripherals.go — the auxiliary-decoder DTOs (ADS-B, M17,
    LoRa/LoRaWAN, APRS, AIS, DSC, MDC1200, paging).

Pure code motion within package config (references resolve across files),
so no behaviour change; only the moved code and the two files' import
sets changed. config.go drops from 3,120 to ~1,972 lines and now holds
just the core schema + loading. go vet + go test ./internal/config green.
